Chore: Refactoring datasets.py step 1a - stimuli tests in own file
